% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJ064_small_last_sce.R
\docType{data}
\name{AJ064_small_last_sce}
\alias{AJ064_small_last_sce}
\title{Small `AJ064` Seurat Data After Processed}
\format{
An object of class \code{Seurat} with 627 rows and 800 columns.
}
\usage{
AJ064_small_last_sce
}
\description{
An object of class Seurat
}
\keyword{datasets}
