% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AJ064_small_sce.R
\docType{data}
\name{AJ064_small_sce}
\alias{AJ064_small_sce}
\title{Small `AJ064` Seurat Data Set}
\format{
An object of class \code{Seurat} with 713 rows and 1000 columns.
}
\usage{
AJ064_small_sce
}
\description{
An object of class Seurat
}
\keyword{datasets}
