% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_filter.R
\name{basic_filter}
\alias{basic_filter}
\title{basic_filter}
\usage{
basic_filter(sce)
}
\arguments{
\item{sce}{An object of class Seurat}
}
\value{
sce.all.filt  An object of class Seurat
}
\description{
filter the genes which show expression less than 3 cells.
filter the cells which percent_mito < 25 & percent_ribo > 3 & percent_hb < 10
filter the cells which  nFeature_RNA > 300 & nFeature_RNA < 8000
}
\examples{
basic_filter(AJ064_small_sce)

}
