% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_markers.R
\name{basic_markers}
\alias{basic_markers}
\title{Basic Markers}
\usage{
basic_markers(sce, org = "human", group = "orig.ident", dir = ".")
}
\arguments{
\item{sce}{An object of class Seurat}

\item{org}{human or mouse, default: human}

\item{group}{default:`orig.ident`, you can change it to `seurat_clusters` or `celltype`}

\item{dir}{the path for saving the figures by `DotPlot` with known famous markers.}
}
\value{
a list of figures by `DotPlot`
}
\description{
Basic Markers
}
\examples{
\donttest{
basic_markers(AJ064_small_last_sce,dir=tempdir())
}
}
