% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_qc.R
\name{basic_qc}
\alias{basic_qc}
\title{Basic Quality Control}
\usage{
basic_qc(sce, org = "human", group = "orig.ident", dir = ".")
}
\arguments{
\item{sce}{An object of class Seurat}

\item{org}{human or mouse, default: human}

\item{group}{default:`orig.ident`,you can change it to `seurat_clusters` or `celltype`}

\item{dir}{the path for saving the figures by `DotPlot` with known famous markers.}
}
\value{
list(p1,p2,p3,sce), the last one in the new `sce`.
}
\description{
add `percent_mito`,`percent_ribo`,`percent_hb` to the Seurat class.
And draw `VlnPlot` for these `qc` values.
}
\examples{
basic_qc(AJ064_small_sce,dir= tempdir())

}
