% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_workflow.R
\name{basic_workflow}
\alias{basic_workflow}
\title{Basic Workflow}
\usage{
basic_workflow(sce, dir = ".")
}
\arguments{
\item{sce}{An object of class Seurat}

\item{dir}{the path for saving the figures by `DotPlot` with known famous markers.}
}
\value{
list(p1,p2,p3,sce), the last one in the new sce with PCA,tSNE,UMAP information.
}
\description{
the workflow from Seurat, including: `NormalizeData`,`FindVariableFeatures`,`ScaleData`,
`RunPCA`,`RunTSNE`,`RunUMAP`,`FindNeighbors`,`FindClusters(sce, resolution = seq(0.1,1,by=0.1))`
we use `clustree` to check the different resolution for `FindClusters`.
}
\examples{
\dontrun{
basic_workflow(AJ064_small_sce,dir=tempdir())
}
}
